requirements.txt
feedparser>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
